v1/bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "A discord bot to fix broken preview such as twitter/x.com...etc with 3rd-party APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.5.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["client*", "config*", "commands*", "url_regex*"]
exclude = ["test*"]